Update on "[CI] Change the benchmark test sharding to a round-robin way"
Summary: This is to alleviate the unbalanced dashboard perf run.

[ghstack-poisoned]
