+++ benchmarks/dynamo/huggingface.py
@@ -366,6 +366,8 @@ class HuggingfaceRunner(BenchmarkRunner):
     def __init__(self):
         super().__init__()
         self.suite_name = "huggingface"
+        model_names = list(BATCH_SIZE_KNOWN_MODELS.keys()) + list(EXTRA_MODELS.keys())
+        self.model_names = sorted(set(model_names))
 
     def load_model(
         self,
@@ -436,24 +438,6 @@ class HuggingfaceRunner(BenchmarkRunner):
 
         self.validate_model(model, example_inputs)
         return device, model_name, model, example_inputs, batch_size
-
-    def iter_model_names(self, args):
-        model_names = list(BATCH_SIZE_KNOWN_MODELS.keys()) + list(EXTRA_MODELS.keys())
-        model_names = set(model_names)
-        model_names = sorted(model_names)
-
-        start, end = self.get_benchmark_indices(len(model_names))
-        for index, model_name in enumerate(model_names):
-            if index < start or index >= end:
-                continue
-            if (
-                not re.search("|".join(args.filter), model_name, re.I)
-                or re.search("|".join(args.exclude), model_name, re.I)
-                or model_name in args.exclude_exact
-                or model_name in SKIP
-            ):
-                continue
-            yield model_name
 
     @property
     def skip_accuracy_checks_large_models_dashboard(self):
