+++ benchmarks/dynamo/timm_models.py
@@ -2,7 +2,6 @@
 import importlib
 import logging
 import os
-import re
 import subprocess
 import sys
 import time
@@ -183,6 +182,7 @@ class TimmRunnner(BenchmarkRunner):
     def __init__(self):
         super().__init__()
         self.suite_name = "timm_models"
+        self.model_name = sorted(TIMM_MODELS.keys())
 
     def load_model(
         self,
@@ -288,23 +288,6 @@ class TimmRunnner(BenchmarkRunner):
         self.validate_model(model, example_inputs)
 
         return device, model_name, model, example_inputs, batch_size
-
-    def iter_model_names(self, args):
-        # for model_name in list_models(pretrained=True, exclude_filters=["*in21k"]):
-        model_names = sorted(TIMM_MODELS.keys())
-        start, end = self.get_benchmark_indices(len(model_names))
-        for index, model_name in enumerate(model_names):
-            if index < start or index >= end:
-                continue
-            if (
-                not re.search("|".join(args.filter), model_name, re.I)
-                or re.search("|".join(args.exclude), model_name, re.I)
-                or model_name in args.exclude_exact
-                or model_name in self.skip_models
-            ):
-                continue
-
-            yield model_name
 
     def pick_grad(self, name, is_training):
         if is_training:
